[refector] test api 호출하는 부분 부모 테스트클래스로 분리
