+++ src/test/java/next/controller/UserAcceptanceTest.java
@@ -8,14 +8,12 @@ import org.junit.jupiter.api.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.test.web.reactive.server.EntityExchangeResult;
-import org.springframework.test.web.reactive.server.WebTestClient;
-import reactor.core.publisher.Mono;
 
 import java.net.URI;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-public class UserAcceptanceTest {
+public class UserAcceptanceTest extends AcceptanceTest {
     private static final Logger logger = LoggerFactory.getLogger(UserAcceptanceTest.class);
 
     @Test
@@ -24,54 +22,22 @@ public class UserAcceptanceTest {
         // 회원가입
         UserCreatedDto expected =
                 new UserCreatedDto("pobi", "password", "포비", "pobi@nextstep.camp");
-        EntityExchangeResult<byte[]> response = client()
-                .post()
-                .uri("/api/users")
-                .body(Mono.just(expected), UserCreatedDto.class)
-                .exchange()
-                .expectStatus().isCreated()
-                .expectBody()
-                .returnResult();
+        EntityExchangeResult<byte[]> response = createResource("/api/users", expected, UserCreatedDto.class);
         URI location = response.getResponseHeaders().getLocation();
         logger.debug("location : {}", location); // /api/users?userId=pobi 와 같은 형태로 반환
 
         // 조회
-        User actual = client()
-                .get()
-                .uri(location.toString())
-                .exchange()
-                .expectStatus().isOk()
-                .expectBody(User.class)
-                .returnResult().getResponseBody();
+        User actual = getResource(location, User.class);
         assertThat(actual.getUserId()).isEqualTo(expected.getUserId());
         assertThat(actual.getName()).isEqualTo(expected.getName());
         assertThat(actual.getEmail()).isEqualTo(expected.getEmail());
 
         // 수정
         UserUpdatedDto updateUser = new UserUpdatedDto("코난", "conan@nextstep.camp");
-        client()
-                .put()
-                .uri(location.toString())
-                .body(Mono.just(updateUser), UserUpdatedDto.class)
-                .exchange()
-                .expectStatus().isOk();
+        updateResource(location, updateUser, UserUpdatedDto.class);
 
-
-        actual = client()
-                .get()
-                .uri(location.toString())
-                .exchange()
-                .expectStatus().isOk()
-                .expectBody(User.class)
-                .returnResult().getResponseBody();
+        actual = getResource(location, User.class);
         assertThat(actual.getName()).isEqualTo(updateUser.getName());
         assertThat(actual.getEmail()).isEqualTo(updateUser.getEmail());
-    }
-
-    private WebTestClient client() {
-        return WebTestClient
-                .bindToServer()
-                .baseUrl("http://localhost:8080")
-                .build();
     }
 }
